+++ client/src/main/java/org/fegin/demo/client/HelloClientApplication.java
@@ -1,39 +1,19 @@
 package org.fegin.demo.client;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import static org.springframework.web.bind.annotation.RequestMethod.GET;
 
 /**
  * @author Spencer Gibb
  */
 @SpringBootApplication
 @EnableDiscoveryClient
-@RestController
 @EnableFeignClients
 public class HelloClientApplication {
-	@Autowired
-	HelloClient client;
-
-	@RequestMapping("/")
-	public String hello() {
-		return client.hello();
-	}
-
 	public static void main(String[] args) {
 		SpringApplication.run(HelloClientApplication.class, args);
 	}
 
-	@FeignClient("HelloServer")
-	interface HelloClient {
-		@RequestMapping(value = "/", method = GET)
-		String hello();
-	}
 }
