+++ server/src/main/java/org/fegin/demo/server/HelloServerApplication.java
@@ -1,33 +1,15 @@
 package org.fegin.demo.server;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.DiscoveryClient;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 /**
  * @author Spencer Gibb
  */
 @SpringBootApplication
 @EnableDiscoveryClient
-@RestController
 public class HelloServerApplication {
-	@Autowired
-	DiscoveryClient client;
-
-	@RequestMapping("/")
-	public String hello() throws JsonProcessingException {
-		List<String> localInstance = client.getServices();
-		return "Hello World: "+ new ObjectMapper().writeValueAsString(localInstance);
-	}
-
 	public static void main(String[] args) {
 		SpringApplication.run(HelloServerApplication.class, args);
 	}
