+++ server/src/main/resources/application.yml
@@ -27,7 +27,7 @@ feign:
   hystrix:
     enabled: true
   httpclient:
-    max-connections: 4
+    max-connections: 1
     max-connections-per-route: 1
     connection-timeout: 2000
   compression:
@@ -35,6 +35,12 @@ feign:
       enabled: true
     response:
       enabled: true
+  client:
+    config:
+      default:
+        connectTimeout: 5000
+        readTimeout: 5000
+        loggerLevel: basic
 
 logging:
   level:
